[Issue-3942] refactor: move interface
